% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciebo_delete.R
\name{sciebo_delete}
\alias{sciebo_delete}
\title{Delete file in Sciebo}
\usage{
sciebo_delete(uni = "uni-bielefeld", user, pw, sciebo_file = NULL)
}
\arguments{
\item{uni}{a string with your university affiliation (e.g., "uni-bielefeld" for Bielefeld University)}

\item{user}{a string with your sciebo user name}

\item{pw}{a string with your sciebo password}

\item{sciebo_file}{a string with the directory and name of the file to be deleted from sciebo ("directory/file.txt")}
}
\description{
Delete file in Sciebo
}
