% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciebo_download.R
\name{sciebo_download}
\alias{sciebo_download}
\title{Downloads file from Sciebo}
\usage{
sciebo_download(
  uni = "uni-bielefeld",
  user,
  pw,
  overwrite = FALSE,
  sciebo_file = NULL,
  local_file = NULL
)
}
\arguments{
\item{uni}{a string with your university affiliation (e.g., "uni-bielefeld" for Bielefeld University)}

\item{user}{a string with your sciebo user name}

\item{pw}{a string with your sciebo password}

\item{overwrite}{a logical indicating whether you want to overwrite existing files with the same name}

\item{sciebo_file}{a string with the directory and name of the file to be downloaded from sciebo ("directory/file.txt")}

\item{local_file}{a string with the directory and name of your local file ("directory/file.txt")}
}
\description{
Downloads file from Sciebo
}
