% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sciebo_upload.R
\name{sciebo_upload}
\alias{sciebo_upload}
\title{Upload file to Sciebo}
\usage{
sciebo_upload(
  uni = "uni-bielefeld",
  user,
  pw,
  overwrite = FALSE,
  sciebo_file,
  local_file
)
}
\arguments{
\item{uni}{a string with your university affiliation (e.g., "uni-bielefeld" for Bielefeld University)}

\item{user}{a string with your sciebo user name}

\item{pw}{a string with your sciebo password}

\item{overwrite}{a logical indicating whether you want to overwrite existing files with the same name}

\item{sciebo_file}{a string with the directory and name of the file to be saved in sciebo ("directory/file.txt")}

\item{local_file}{a string with the directory and name of your local file ("directory/file.txt")}
}
\description{
Upload file to Sciebo
}
